requirements.txt
asgiref==3.6.0
attrs==22.2.0
black==23.1.0
click==8.1.3
Django==4.1.6
djangorestframework==3.14.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pluggy==1.0.0
pycodestyle==2.10.0
pytest==7.2.1
pytest-django==4.5.2
pytest-testdox==3.0.1
pytz==2022.7.1
sqlparse==0.4.3
